+++ src/valvis/export/VisitRef.js
@@ -2,32 +2,56 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { BaseDefRef } from '#x/BaseDefRef';
+import { VisitDef } from '#x/VisitDef';
 
 
 /**
- * Forward declaration of this class, because `import`ing it would cause a
- * circular dependency while loading.
- *
- * @typedef BaseValueVisitor
- * @type {object}
- */
-
-/**
- * Companion class of {@link BaseValueVisitor}, which represents the result of a
- * visit of a value that had been visited elsewhere during a visit.
+ * Representative of the result of a visit of a value that had been visited
+ * elsewhere during a visit.
  *
  * Along with just having a record of the shared nature of the structure,
  * instances of this class are also instrucmental in "breaking" circular
  * references during visits, making it possible to fully visit values that have
- * such circular references. See {@link BaseValueVisitor#_impl_shouldRef} for
- * more details.
+ * such circular references.
  */
 export class VisitRef extends BaseDefRef {
-  // @defaultConstructor
+  /**
+   * This instance's corresponding def.
+   *
+   * @type {VisitDef}
+   */
+  #def;
+
+  /**
+   * Constructs an instance.
+   *
+   * @param {VisitDef} def The corresponding def.
+   */
+  constructor(def) {
+    const valueArg = def.isFinished() ? [def.value] : [];
+    super(def.index, ...valueArg);
+
+    this.#def = def;
+  }
+
+  /** @override */
+  get def() {
+    return this.#def;
+  }
 
   /** @override */
   get ref() {
     return this;
+  }
+
+  /** @override */
+  get value() {
+    return this.#def.value;
+  }
+
+  /** @override */
+  isFinished() {
+    return this.#def.isFinished();
   }
 
   /**