+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Breaking changes:
   * Renamed some methods in `BaseValueVisitor`, for clarity and consistency, and
     introduced a public method `visitWrap()`, which has the same "synchronous if
     possible but async if not" behavior that is used internally.
+  * Reworked `VisitDef` and `VisitRef` to not assume an associated visitor
+    instance.
 
 Other notable changes:
 * None.